requirements.txt
pymupdf
paddlepaddle==2.4.2
paddleocr
langchain==0.0.146
transformers==4.27.1
unstructured[local-inference]
layoutparser[layoutmodels,tesseract]
nltk
sentence-transformers
beautifulsoup4
icetk
cpm_kernels
faiss-cpu
accelerate
gradio==3.28.3
fastapi
uvicorn
peft
pypinyin
bitsandbytes
click~=8.1.3
tabulate
wudao
python-dotenv
